requirements.txt
altair==4.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blessings==1.7
blinker==1.4
bqplot==0.12.33
branca==0.5.0
cachetools==5.2.0
certifi
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
colorcet==3.0.0
colour==0.1.5
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.1
decorator==5.1.1
defusedxml==0.7.1
earthengine-api==0.1.316
ee==0.2
ee-extra==0.0.13
entrypoints==0.4
fastjsonschema==2.15.3
ffmpeg-python==0.2.0
filelock==3.7.1
folium==0.12.1.post1
fonttools==4.34.2
future==0.18.2
gdown==4.5.1
geeadd==0.5.5
geemap==0.14.3
geetools==0.6.14
geocoder==1.38.1
geojson==2.5.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
httplib2==0.20.4
httplib2shim==0.0.3
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.15.0
ipyleaflet==0.17.0
ipython==7.34.0
ipython-genutils==0.2.0
ipytree==0.2.1
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.8
jsonschema==4.6.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
logzero==1.7.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.4.2
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
param==1.12.2
parso==0.8.3
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.9.0
plotly-express==0.4.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyCRS==1.0.2
pyct==0.4.8
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
pyshp==2.3.0
PySocks==1.7.1
python-box==6.0.2
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.6
pyzmq==23.2.0
ratelim==0.1.6
requests==2.28.1
rich==12.4.4
rsa==4.8
sankee==0.1.0
scipy==1.7.3
scooby==0.5.12
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
statsmodels==0.13.2
streamlit==1.10.0
streamlit-folium==0.6.14
tenacity==8.0.1
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
traittypes==0.2.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uritemplate==3.0.1
urllib3==1.26.9
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
whitebox==2.1.2
whiteboxgui==0.7.0
widgetsnbextension==3.6.1
wincertstore==0.2
xyzservices==2022.6.0
zipp==3.8.0
